move catch all to app/page
